requirements.txt
requests>=2.31.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2 